remove unneeded file
