+++ server/service/testing_utils.go
@@ -16,6 +16,7 @@ import (
 	"net/http/httptest"
 	"os"
 	"sort"
+	"strings"
 	"sync"
 	"testing"
 	"time"
@@ -1540,4 +1541,25 @@ type acmeCSRSigner struct {
 
 func (a *acmeCSRSigner) SignCSR(_ context.Context, csr *x509.CertificateRequest) (*x509.Certificate, error) {
 	return a.signer.Signx509CSR(csr)
+}
+
+// mockRoundTripper is a custom http.RoundTripper that redirects requests to a mock server.
+type mockRoundTripper struct {
+	mockServer  string
+	origBaseURL string
+	next        http.RoundTripper
+}
+
+func (rt *mockRoundTripper) RoundTrip(req *http.Request) (*http.Response, error) {
+	if strings.Contains(req.URL.String(), rt.origBaseURL) {
+		path := strings.TrimPrefix(req.URL.Path, "/")
+		newURL := fmt.Sprintf("%s/%s", rt.mockServer, path)
+		newReq, err := http.NewRequestWithContext(req.Context(), req.Method, newURL, req.Body) //nolint:gosec // test helper, URL is from mock server
+		if err != nil {
+			return nil, err
+		}
+		newReq.Header = req.Header
+		return rt.next.RoundTrip(newReq)
+	}
+	return rt.next.RoundTrip(req)
 }
